% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HospitalCensusApp-package.R
\docType{package}
\name{HospitalCensusApp-package}
\alias{HospitalCensusApp-package}
\alias{_PACKAGE}
\title{HospitalCensusApp: Transplant Census Prediction Application}
\description{
A shiny application for prediction of transplant census based on user inputs.
}
\author{
\strong{Maintainer}: Ezra Porter \email{porterej@chop.edu}

}
\keyword{internal}
